backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.0

# Core
pydantic>=1.10,<2.0
httpx>=0.27,<0.29
typer>=0.7,<0.10

# ML stack
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scipy>=1.11,<2.0
scikit-learn>=1.4,<1.5

# XAI + ML
xgboost>=2.0,<2.1
shap>=0.45,<0.46

# NLP stack (STRICT group)
spacy==3.7.4
thinc==8.2.5
blis==0.7.11
srsly>=2.5,<3.0
preshed>=3.0,<4.0
cymem>=2.0,<3.0
murmurhash>=1.0,<2.0

# spacy model
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl

# numba stack (STRICT)
numba==0.64.0
llvmlite==0.46.0

# Database
sqlalchemy>=2.0,<2.1
aiosqlite>=0.19,<1.0
asyncpg>=0.29,<1.0
greenlet>=3.0,<4.0

# misc
requests>=2.32,<3
python-dateutil>=2.9,<3
typing-extensions>=4.10,<5